Add nascent CONTROL_DIR: for sans-API operation
This Poller implementation looks at a "control" directory periodically
instead of the Heroku API.  Right now, it only works with the
SimpleDynoDriver, because the DockerDynoDriver is going to have to be
taught how cope with the slug source being on a local file system.

The invocation looks a bit like this:

    CONTROL_DIR=a_dir ./hsup run -a app -- /bin/sh -c '/usr/bin/printenv'

Normally, that directory will contain documents ("new", or "loaded")
that look something like this:

    {
        "Version": 1,
        "Env": {
            "NAME": "CONTENTS"
        },
        "Slug": "sample-slug.tgz",
        "Processes": [
            {
                "Args": ["./web-server", "arg"],
    	    "Quantity": 2,
    	    "Type": "web"
            },
            {
                "Args": ["./worker", "arg"],
    	    "Quantity": 2,
    	    "Type": "worker"
            }
        ]
    }

In that case, text from printenv as the following would be returned:

    NAME=CONTENTS

The control directory looks like this:

    name/loaded
    name/new
    name/rejected
    name/last_error

To trigger a new release, write to the file "new".  Periodically, the
existence of such a file is checked.  If it loads successfully, it is
moved to (and can replace) the "loaded" file.

If validating a "new" file fails, the file is moved to the "rejected"
file, and the "last_error" file is written with a human-readable reason
for why.

The "loaded" file is useful to be able to inspect a successfully loaded
state and as a sanity check: it should always be a valid document to
load when the program starts.

This scheme is heavily borrowed (in implementation and conception) from
pg_logplexcollector, which has used it for hot-loading new
configurations for quite a while now smoothly.  That is why this code is
overwrought relative to the rest of "hsup".