Add abspath driver
This can act as a supervisor inside a chroot or container environment,
where absolute paths can be assumed to work more like Heroku Cedar.
Otherwise, it is much like the "simple" dyno driver.

Because this driver is somewhat more tedious to experiment with, add
documentation (to WIP.md) and the "abspath-model" tool.

This is not portable: Linux is required.

Accolades: Dan Peterson <dan@heroku.com>, whose dyno-run program was
inspected to concoct this with his permission.