Simplify abspath driver: no priv dropping
It was thought the abspath driver could learn to not run as root, but
Go's potentially pervasive use of threads has made it impossible to use
the "setuid" and "setgid" system calls reliably.  This has culminated in
the effective removal of those system calls for now:

    https://code.google.com/p/go/issues/detail?id=1435

Instead, use a program like setuidgid to drop privileges.  This is
reflected in WIP.md.

Also remove the "base" setting which is superfluous.  The onus is again
on the caller to call "chroot," for instance.