Use abspath "build" for docker dyno driver
Instead of the "curl" expression, push "hsup" into the image to fetch
the slug and build the image.  With more work, the "control directory"
format of hsup can be used, so that the build process inside the
container does not require having to talk to the Heroku API.

That opens the door for alternate slug sources, e.g. local builds while
developing.

Also fix up running the slug-fetching process as root: instead, employ
setuidgid to drop privileges.