Improve "start" command and change its defaults
"hsup start" now accepts a specification like this:

    web=1 worker=2 ....

And runs the specified formation type with the specified concurrency.

Without a specification, it will run, with concurrency=1, any process
that has a scale factor above zero.  So, for example:

    hsup start -a myapp

Might run "web" but not the "console" formation in the case of Heroku
Ruby applications.