Add and use gob poller for self-invocations
This piggybacks on the CONTROL_DIR json-demarshalling structures to
serialize a gob passed by environment, allowing hsup to invoke itself
inside containers in both build and execution phases with rich
information and through a single mechanism, cleaning up all the
inspection of how hsup was configured when pushing data into the
sub-invocation.

Unlike the other pollers, the gob poller will only ever send exactly one
"Processes" value.

Detailed design information follows:

While this patch solves more problems than it creates, it exposes and
exacerbates some cracks in the software engineering:

Firstly, the guts of one poller (the "ControlDir") is shared between two
pollers, one of which has nothing to do with a directories.  To make
matters worse, its serialization/deserialization is in the guts of a
DynoDriver.

Superficially, this is an alarming destruction of abstraction barriers,
but not so much in light of what a "ControlDir" is at a detailed level:
to be a compact, no-frills, serializable data structure describing all
the same data any poller needs to respond.  It is the concrete variant
of the interface-driven approach used to return data from pollers
thusfar (e.g. the "Formation" interface).

In light of how useful serializing rich values to instruct descended
hsup processes is, perhaps all pollers ought to be adjusted to yield a
single, concrete implementation (that would be exactly like
"ControlDir") rather than support any kind of polymorphism via
interfaces as is currently the case.

This organizational and naming snarl is worthwhile removing if this
patch survives service to a larger goal: making the CONTROL_DIR
functionality work, and in particular supporting both http and
file-driven image building.  Before, this was busted and it appeared it
some futzing around with volumes and binds to fix it, but I found this
harder than anticipated when trying to achieve it.

On this unexpected difficulty: I chose this design after experimenting
with another.  That approach was using the CONTROL_DIR format to teach
the hsup self-invocation what to do.  Implied in this is file
generation, manipulation, and particularly sharing, and in this are the
seeds of that approach's undoing:

* There a lot more that can go wrong that requires error handling in
  code paths where one can do without such failure modes.

* Dockerfiles don't support bind mounting from the host, presumably as
  it defeats the portability that is a goal of that format: one would
  have to write a bunch of code to start a series of containers with
  bind mounts to the host.

* One needs to dispose of temporary directories when an Executor is
  completely done with. So one needs to bloat a new operator to dispose
  of it, as well as fields to track what to clean up.  And, that
  clean-up has failure modes, too.

* The CONTROL_DIR hot loading support needs to move and create files,
  and file UID/GID mapping through bind mounts in other containers is
  complex.  It gets worse when one considers how boot2docker does file
  sharing on Macintosh: then *three* possible interpretations or
  mangling of file UIDs come into play.

Finally, the useful part about CONTROL_DIR -- the ability to poll and
hot-load the contents -- is not employed at all inside a container,
because a new release triggers a start of a new container, and that's a
perfect opportunity to inject the immutable HSUP_CONTROL_GOB.  This
fastidiousness is of little consequence if the system degenerated well
otherwise, but it did not.  The HSUP_CONTROL_GOB style, sacrificing the
never-used hot-updates, eliminates a ton of complexity, and teaches an
object lesson: do not underestimate file management and containers.