Add support for local tarballs
This is done by copying them (nested) into the generated tarball
submitted along with the Dockerfile along with image build.

To support this, introduce the notion of "SlugWhere" which is used to
detect where the slug is being referenced is coming from.  HTTP (which
includes HTTPS) and Local are the supported options here.