Fix the confused organization of GobNotifier
In 06571ad65632ebcbfeb497223adc40cdb16e1ff5, there is some commentary
made as to how confused the code placement is with regard to teaching
hsup how to self-invoke with a gob.

Now that local slugs are proven to work with this model, fix the code
organization: there is now a "serialization" file that contains
structures that have the express purpose of being Marshal-able with JSON
or Gob that are shared by both implementations and can be less
(superficially) alarmingly be called deep on the DynoDrivers.