Fix bogus use of Stop in Executor
The "Restart" input while "Started" issues a "Stop()" but then attempts
an epsilon-move via "again", which is of no use because the executor
state has not been modified.

It would appear this would get into a tight loop until "Stop()" finally
returns an error, letting the state machine process the "Exited" input
that arrives from the "Wait()" implied in starting any process.

Similarly, in the Retiring state, the success of "Stop()" is used to
retire the state machine.  However, in event of some races, it's quite
possible the process has died already and "Stop()" will error
continuously, never allowing progression to "Retired".  Similar to the
approach above, rely on the "Exited" message which is thought to be
forthcoming from the "Wait()" implied by "Start()".