Move hsup invocation info to a serializable struct
With this patch, all the information necessary to sub-invoke hsup is
captured in a single, serializable struct.  The GobNotifier has been
adjusted accordingly, and the sub-invocations no longer require (and are
discouraged from using) any arguments.  Command line parsing also yields
this representation to the rest of the hsup package.

One of removed features in this is the concurrency option when using
"hsup run".  This was not commonly used and fits in poorly with the
dominant "start" mode.  It could be reintroduced without much penalty if
necessary.

Overall, there is some penalty to line count, but a number of these are
from documentation.  Otherwise, the density and mysteriousness of the
lines sub-invocation hsup are now less cryptic and lighter on the page.