Add a check for stack compatibility
Augment the API poller to relay the stack name, and then check for the
stack in docker_dyno_driver.

This was instigated by blindly attempting to run a Cedar app as
Cedar-14, which failed in a cryptic way, e.g. linking errors against
OpenSSL.  This was opaque in identifying the cause of the fault, so this
patch makes the matter explicit in hsup instead.