Implement diag in-situ debug log
Diag is a package -- employed mostly for a convenient global variable,
ala "log" -- to support in-situ debugging.  A "diag" dump can be
retrieved from "hsup" by sending "SIGUSR1".

This allows the retirement of some of the noisier diagnostic output
still present in hsup, and the enrichment and agglomeration of more
useful traces.