`bash -c` requires a single word as its argument
bash word splitting on the profile runner (abspath driver) was
preventing containers with commands split into multiple entries (eg.:
`Args: ["echo", "$PORT", "&&", "echo", "$DYNO"]`) from executing.

$@ has special treatment by bash and by default will always be subject
to word splitting. This patch is a hack to avoid it being automatically
split into multiple words, as `bash -c` requires a single word as its
argument.

In the future we may consider switching the public interface of `Args`
to be an opaque string to be interpreted by `bash`, instead of an array
of words to be passed to `execve(2)`. This works better with `hsup run`
too, as we don't need to bother splitting words and let `bash -c` handle
it.

/cc @fdr