IP allocation for the libcontainer driver
Rely on unique UIDs per container to pick a private subnet for them.
Subnets are of size /30 (the smallest possible subnet), allocated from
the 172.16/12 block (private address space as per RFC1918).

This tries to avoid allocating IPs known to be used by AWS on EC2
classic (e.g.: their internal DNS server is at 172.16.0.23).

What /30 subnets to use is currently hardcoded, but this can possibly be
changed in the future to be configurable by driver parameters.