valid exit status when killed by signal
It's standard practice (e.g.: bash does it) to exit with status code 128+signal
when terminated by a unix signal.

Also add a note (TODO to be fixed later) that hsup itself being signalled will
force the exit code to be 1 with os.Exit(1), regardless of what are the exit
status of processes it supervises.