proper libcontainer.Stop() implementation
namespaces.Exec already does the necessary work waiting for the supervised
process to exit. The libcontainer driver just needs to capture the exit code
returned by namespaces.Exec and propagate it back to driver.Wait().

driver.Stop() just needs to send a SIGTERM to the underlying driver (abspath)
running inside the container, so it can handle graceful shutdown. There is a
rare race condition when Stop() is called before the libcontainer-init driver
re-execs itself as abspath inside the container, in which case the
libcontainer-init driver will ignore the SIGTERM and processes will not be
terminated.

The race condition will be handled in future patches, but it should not be a big
deal for now.