normalize IP and port assignment across drivers
Ensure that ENV["PORT"] is set for all apps before driver.Start() gets
called, so drivers can rely on that to determine which port dynos will
bind to.

Avoid re-calculating ports and IP addresses in cases when drivers can
cache it for a dyno (Executor.IPInfo).

Remove the direct dependency of the ControlAPI on drivers. Instead it
just depends on the Executor. Drivers are responsible for filling
information inside the Executor.

Docker containers with the docker driver where taking a little over than
10s to start on my machine, so I bumped test timeouts to ~30s.